requirement.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-crispy-forms==1.13.0
django-environ==0.7.0
django-extensions==3.1.3
django-filter==21.1
django-pydenticon==0.2
django-taggit==1.5.1
djangorestframework==3.12.4
idna==3.2
libgravatar==1.0.0
Markdown==3.3.4
mysqlclient==2.0.3
oauthlib==3.1.1
Pillow==8.3.2
pycparser==2.20
pydenticon==0.3.1
PyJWT==2.1.0
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
urllib3==1.26.7
